+++ src/__tests__/server/default/schema.test.mjs
@@ -43,6 +43,17 @@ const requiredTypes = [
       ],
     },
   },
+  {
+    name: "UserCreateInput",
+    fields: {
+      both: [
+      ],
+      prisma: [
+      ],
+      api: [
+      ],
+    },
+  },
 ]