+++ src/server/index.mjs
@@ -1,7 +1,6 @@
 
 import startServer from "@prisma-cms/server";
 import CoreModule from "./modules";
-import { ImagesMiddleware } from "@prisma-cms/upload-module";
 
 
 const coreModule = new CoreModule({
@@ -9,13 +8,10 @@ const coreModule = new CoreModule({
 
 const resolvers = coreModule.getResolvers();
 
-const imagesMiddleware = new ImagesMiddleware().processRequest;
-
 
 startServer({
   typeDefs: 'src/schema/generated/api.graphql',
   resolvers,
-  imagesMiddleware,
 });