+++ src/server/scripts/deploy/index.mjs
@@ -2,10 +2,12 @@
 import {
   buildApiSchema,
   deploy,
+  buildApiFragments,
 } from "@prisma-cms/server/src/server/scripts/deploy/";
 
 
 export {
   buildApiSchema,
   deploy,
+  buildApiFragments,
 }