+++ src/components/Renderer/pages/society/ChatRooms/index.js
@@ -1,0 +1,34 @@
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+
+import Page from "../../layout";
+
+import {
+  ChatRooms,
+} from "@prisma-cms/society";
+
+
+class ChatRoomsPage extends Page {
+
+
+  setPageMeta() {
+
+    return super.setPageMeta({
+      title: "Чаты",
+    });
+
+  }
+
+
+  render() {
+
+    return super.render(
+      <ChatRooms
+        {...this.props}
+      />
+    );
+  }
+}
+
+
+export default ChatRoomsPage;